Create metadata extraction module for Spring Content.
Split metadata extraction from renditions into a dedicated module, `spring-content-metadata-extraction`, with its own configurations, documentation, and Maven artifacts. Updated package structure, dependencies, and imports accordingly.
