+++ pom.xml
@@ -113,6 +113,8 @@ Integrates with Spring Data, Spring Data REST and Apache Solr</description>
         <module>./spring-content-s3-boot-starter</module>
         <module>./spring-content-renditions</module>
         <module>./spring-content-renditions-boot-starter</module>
+        <module>./spring-content-metadata-extraction</module>
+        <module>./spring-content-metadata-extraction-boot-starter</module>
         <module>./spring-content-solr</module>
         <module>./spring-content-solr-boot-starter</module>
         <module>./spring-content-elasticsearch</module>
