+++ spring-content-metadata-extraction-boot-starter/pom.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>spring-content</artifactId>
+        <groupId>com.github.paulcwarren</groupId>
+        <version>3.0.17-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>spring-content-metadata-extraction-boot-starter</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.github.paulcwarren</groupId>
+            <artifactId>spring-content-metadata-extraction</artifactId>
+            <version>3.0.17-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.github.paulcwarren</groupId>
+            <artifactId>spring-content-autoconfigure</artifactId>
+            <version>3.0.17-SNAPSHOT</version>
+        </dependency>
+    </dependencies>
+</project>
