+++ spring-content-metadata-extraction-boot-starter/src/main/resources/spring.provides
@@ -1,0 +1,1 @@
+provides: spring-content-metadata-extraction
