+++ spring-content-metadata-extraction/src/main/java/org/springframework/content/metadataextraction/MetadataExtractionException.java
@@ -14,7 +14,7 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
-package org.springframework.content.renditions.metadataextractors;
+package org.springframework.content.metadataextraction;
 
 import org.springframework.content.commons.renditions.MetadataExtractor;
