+++ spring-content-metadata-extraction/src/test/java/org/springframework/content/metadataextraction/extractors/DefaultMetadataExtractorTest.java
@@ -14,9 +14,10 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
-package org.springframework.content.renditions.metadataextractors;
+package org.springframework.content.metadataextraction.extractors;
 
 import org.junit.Test;
+import org.springframework.content.metadataextraction.MetadataExtractionException;
 
 import java.io.File;
 import java.io.IOException;
